Vit_s16 model change
